requirements.txt
altair==4.1.0
astor==0.8.1
atomicwrites==1.4.0
attrs==19.3.0
backcall==0.1.0
base58==2.0.1
bleach==3.1.5
blinker==1.4
blis==0.4.1
boto==2.49.0
boto3==1.13.11
botocore==1.16.11
cachetools==4.1.0
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
cymem==2.0.3
Cython==0.29.14
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
editdistance==0.5.3
en-core-web-sm==2.2.5
entrypoints==0.3
enum-compat==0.0.3
fastparquet==0.4.0
funcy==1.14
future==0.18.2
fuzzywuzzy==0.18.0
gensim==3.8.3
idna==2.9
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
langdetect==1.0.8
llvmlite==0.32.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mock==4.0.2
more-itertools==8.3.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.6
nltk==3.5
notebook==6.0.3
numba==0.49.1
numexpr==2.7.1
numpy==1.18.4
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
patsy==0.5.1
pickleshare==0.7.5
Pillow==7.1.2
plac==1.1.3
pluggy==0.13.1
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.12.2
py==1.8.1
pydeck==0.4.0b2
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.2
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2020.1
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
regex==2020.5.14
requests==2.23.0
s3transfer==0.3.3
scikit-learn==0.23.0
scikit-surprise==1.1.1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
smart-open==2.0.0
spacy==2.2.4
srsly==1.0.2
statsmodels==0.11.1
streamlit==0.61.0
surprise==0.1
terminado==0.8.3
testpath==0.4.4
thinc==7.4.0
threadpoolctl==2.0.0
thrift==0.13.0
toml==0.10.1
toolz==0.10.0
tornado==5.1.1
tqdm==4.46.0
traitlets==4.3.3
tzlocal==2.1
urllib3==1.25.9
validators==0.15.0
wasabi==0.6.0
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==1.2.0